% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_cbdd.r
\name{cbdd.fromRCX}
\alias{cbdd.fromRCX}
\title{Create CBDD-like data from \code{\link{RCX}} object}
\usage{
cbdd.fromRCX(rcx, useNamespace = NULL)
}
\arguments{
\item{rcx}{\code{\link{RCX}} object}

\item{useNamespace}{string (optional) name of the namespace for nodes}
}
\value{
data.frame with two columns containing IDs of interacting nodes. Returns NULL if rcx object does not contain edge information.
}
\description{
Returns CBDD-like data as defined at https://cbdd.thomsonreuterslifesciences.com/cbdd/help_algorithms/
The identifiers used are (in this order, if applicable) the namespace ids, the "r"epresents field, the "n"ame field or the @id field of the CX data
}
\examples{
## Establish a server connection
ndexcon = ndex.connect()
## Find a network and get its UUID
networks = ndex.find.networks(ndexcon,"p53")
networkId = networks[1,"externalId"]
## Get the network data 
rcx = ndex.get.network(ndexcon, networkId) 
## Convert to cbbd
cbdd = cbdd.fromRCX(rcx)
}
