% Generated by roxygen2 (4.0.2): do not edit by hand
\name{df2properties}
\alias{df2properties}
\title{Convert data frame of properties back to the list structure eligible to be used with propertyGraph}
\usage{
df2properties(df, missing.value = NA)
}
\arguments{
\item{df}{data frame made of node or edge properties}
}
\value{
properties in the list form
}
\description{
Convert data frame of properties back to the list structure eligible to be used with propertyGraph
}
\details{
This is needed to save the networks back to NDEx
}

