% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_connect.r
\name{get.network.api}
\alias{get.network.api}
\title{List possible queries to NDEx API}
\usage{
get.network.api(ndexcon)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}
}
\value{
data.frame detaling the API names, paths, parameters and athentication needed for API requests.
}
\description{
List possible queries to NDEx API
}
\details{
Retrieves information on the NDEx API calls
}
\examples{
##TBD
}
\seealso{
\code{\link{ndex_rest_GET}},  \code{\link{ndex_rest_PUT}},  \code{\link{ndex_rest_POST}}
}

