% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_maintenance.R
\name{listToYaml}
\alias{listToYaml}
\title{Convert an R object into a YAML string
same as yaml::as.yaml(obj)
only for package maintenance!}
\usage{
listToYaml(obj, indent = "  ", indentShift = "")
}
\arguments{
\item{obj}{list; Nested list to be translated. The allowed list elements are: list, character, numeric}

\item{indent}{character (optional); Character(s) used for the indentation. Default: <tab>}

\item{indentShift}{character (optional); Shifts the block to the right by putting the characters before each line (mostly used in internal recursion)}
}
\value{
character; R code for generating the nested list
}
\description{
Convert an R object into a YAML string
same as yaml::as.yaml(obj)
only for package maintenance!
}
\examples{
\dontrun{test = list(bla='some text',blubb=list(a='more text', version='2.0'),justANumber=123456)
#$bla
#[1] "some text"
#$blubb
#$blubb$a
#[1] "more text"
#$blubb$version
#[1] "2.0"
#$justANumber
#[1] 123456

listToYaml(test)
#[1] "bla: some text\\nblubb:\\n  a: more text\\n  version: '2.0'\\njustANumber: 123456.0\\n"
}
}
