% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_connect.r
\name{ndex.alive}
\alias{ndex.alive}
\title{Check if user is authenticated to NDEx REST server}
\usage{
ndex.alive(ndexcon)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}
}
\value{
logical (TRUE if user is authenticated and connection is active, FALSE otherwise)
}
\description{
Check if user is authenticated to NDEx REST server
}

