% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_maintenance.R
\docType{data}
\name{ndex.api.conf.header}
\alias{ndex.api.conf.header}
\title{Default header for the ndex.api.conf.r file
only for package maintenance!}
\format{An object of class \code{character} of length 1.}
\usage{
ndex.api.conf.header
}
\description{
Default header for the ndex.api.conf.r file
only for package maintenance!
}
\keyword{datasets}

