% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_api_conf.r, R/ndex_api_conf_test.r
\docType{data}
\name{ndex.api.conf}
\alias{ndex.api.conf}
\title{NDEx server api configuration}
\format{An object of class \code{list} of length 2.}
\usage{
ndex.api.conf

ndex.api.conf
}
\value{
Nested list resembling the NDEx server REST API structure

Nested list resembling the REST structure
}
\description{
This nested list contains the url and methods for accessing the NDEx server via its REST full api
It contains specifications for NDEx server api version 1.3 and 2.0
If possible, the version 2.0 should be used

This nested list contains the url and methods for accessing the NDEx server via its REST full api
It contains specifications for NDEx server api version 1.3 and 2.0
If possible, the version 2.0 should be used
}
\keyword{datasets}

