% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_api_conf.r
\docType{data}
\name{ndex.api}
\alias{ndex.api}
\title{NDEx server api configuration}
\format{An object of class \code{list} of length 3.}
\usage{
ndex.api
}
\value{
Nested list resembling the REST structure
}
\description{
This nested list contains the url and methods for accessing the NDEx server via its REST full api
It contains specifications for NDEx server api version 1.3 and 2.0
If possible, the version 2.0 should be used
}
\keyword{datasets}

