% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_maintenance.R
\docType{data}
\name{ndex.conf.header}
\alias{ndex.conf.header}
\title{Default header for the ndex.conf.r file}
\format{An object of class \code{character} of length 1.}
\usage{
ndex.conf.header
}
\description{
Default header for the ndex.conf.r file
}
\note{
only for package maintenance!
}
\examples{
ndex.conf.header
}
\keyword{datasets}
