% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks.r
\name{ndex.create.network}
\alias{ndex.create.network}
\title{Create a Network from CX data}
\usage{
ndex.create.network(ndexcon, rcx)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{rcx}{\code{\link{RCX}} object}
}
\value{
\code{\link{RCX}} object
}
\description{
POST : /network/asCX
TODO: currently doesnt seem to be working.Debug!
This method creates a new network on the NDEx server.
}
\section{REST query}{

POST : /network/asCX
}

