% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks_provenance.R
\name{ndex.create.provenance.entity}
\alias{ndex.create.provenance.entity}
\title{Create Network Provenance Entity}
\usage{
ndex.create.provenance.entity(uri = NA, creationEvent, property.name,
  property.value)
}
\arguments{
\item{uri}{URI of the resource described by the ProvenanceEntity. This field will not be set in some cases, such as a file upload or an algorithmic event that generates a network without a prior network as input.}

\item{creationEvent}{ProvenanceEvent - Has semantics of PROV:wasGeneratedBy}

\item{property.name}{array of names for name-value pairs}

\item{property.value}{array of values for name-value pairs}
}
\value{
Network ProvenanceEntity Object
}
\description{
Create Network Provenance Entity
}
\examples{
\dontrun{}
}
