% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks_provenance.R
\name{ndex.create.provenance.entity.upload}
\alias{ndex.create.provenance.entity.upload}
\title{Create a default Network ProvenanceEntity object for upload}
\usage{
ndex.create.provenance.entity.upload()
}
\value{
Network ProvenanceEntity Object
}
\description{
Create a default Network ProvenanceEntity object for upload
}
\examples{
\dontrun{}
}
