% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks_provenance.R
\name{ndex.create.provenance.event}
\alias{ndex.create.provenance.event}
\title{Create Network Provenance Event}
\usage{
ndex.create.provenance.event(inputs = NA, startedAtTime, endedAtTime,
  eventType, property.name, property.value)
}
\arguments{
\item{inputs}{array of ProvenanceEntity objects - Has semantics of PROV:used}

\item{startedAtTime}{timestamp - Has semantics of PROV:startingAtTime}

\item{endedAtTime}{timestamp - Has semantics of PROV:endingAtTime}

\item{eventType}{name or description of the event (e.g. "File Upload")}

\item{property.name}{array of names for name-value pairs}

\item{property.value}{array of values for name-value pairs}
}
\value{
Network ProvenanceEvent Object
}
\description{
Create Network Provenance Event
}
\examples{
\dontrun{}
}
