% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_groups.R
\name{ndex.delete.group}
\alias{ndex.delete.group}
\title{Delete Group}
\usage{
ndex.delete.group(ndexcon, groupId)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{groupId}{character; unique ID (UUID) of the group}
}
\value{
NULL if successfull, else an error is thrown
}
\description{
Delete the group specified by groupId
}
\note{
Compatible to NDEx server version 2.0
}
\section{REST query}{

DELETE: ndex.api.config$api$group$delete
}

\examples{
\dontrun{
groupId = "aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee"
ndexr.delete.group(ndexcon,groupId)
}
}
