% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks_simple_operations.r
\name{ndex.delete.network}
\alias{ndex.delete.network}
\title{Delete a network}
\usage{
ndex.delete.network(ndexcon, nuuid)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{network_id}{UUID of the network}
}
\description{
Delete a network
}
\section{REST query}{

}

