% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks.r
\name{ndex.find.networks}
\alias{ndex.find.networks}
\title{Search networks in NDEx (by description)
    network    POST    /network/search/{skipBlocks}/{blockSize}    SimpleNetworkQuery    NetworkSummary[]}
\usage{
ndex.find.networks(ndexcon, searchString = "", accountName, skipBlocks = 0,
  blockSize = 10)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{searchString}{string by which to search}

\item{accountName}{string; constrain search to networks administered by this account}

\item{skipBlocks}{-}

\item{blockSize}{-}
}
\value{
Data frame with network information: ID, name, whether it is public, edge and node count; source and format of network. NULL if no networks are found.
}
\description{
Search networks in NDEx (by description)
    network    POST    /network/search/{skipBlocks}/{blockSize}    SimpleNetworkQuery    NetworkSummary[]
}
\note{
Search strings may be structured
}
\examples{
\dontrun{ndex.find.networks("p53")}
}

