% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_users.R
\name{ndex.find.user.byId}
\alias{ndex.find.user.byId}
\title{Get User By UUID}
\usage{
ndex.find.user.byId(ndexcon, uuuid)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{uuuid}{unique ID of the user}
}
\value{
list of properties describing the user (externalId, emailAddress, website, etc.). Throws error (404) if user isn't found!
}
\description{
Get User By UUID
}
\note{
Compatible to NDEx server version 2.0
}
\section{REST query}{

GET: ndex.api.config$api$user$get$byId
}

\examples{
\dontrun{
user = ndex.find.user.byName(ndexcon, 'someAccountName')
user = ndex.find.user.byId(ndexcon, user$externalId)
}
}
