% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_users.R
\name{ndex.find.user.byName}
\alias{ndex.find.user.byName}
\title{Get User By Name}
\usage{
ndex.find.user.byName(ndexcon, name)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{name}{name of the user}
}
\value{
list of properties describing the user (externalId, emailAddress, website, etc.). Throws error (404) if user isn't found!
}
\description{
Get User By Name
}
\note{
Compatible to NDEx server version 2.0
}
\section{REST query}{

GET: ndex.api.config$api$user$get$byName
}

\examples{
\dontrun{
user = ndex.find.user.byName(ndexcon, 'someAccountName')
}
}
