% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ndex.get.complete.network.as.property.graph}
\alias{ndex.get.complete.network.as.property.graph}
\title{Get complete network as property graph}
\usage{
ndex.get.complete.network.as.property.graph(network_id)
}
\arguments{
\item{network_id}{unique ID of the network}
}
\value{
\code{\link{ndexgraph}} object
}
\description{
Get complete network as property graph
}
\details{
Uses getEdges (this procedure will return complete network with all elements)
Nodes use primary ID of the base term ('represents' element)
Edges use primary ID of the base term ('predicate', or 'p' element)
Mapping table for the nodes is retrieved ('alias' and 'related' terms) to facilitate conversions/data mapping
}
\section{REST query}{

Executes GET query /network/{networkUUID}/asPropertyGraph, returns parsed PropertyGraphNetwork object
}
\examples{
\dontrun{
x <- ndex.get.complete.network.as.property.graph('32733ca9-4e84-11e4-98fb-000c29873918')
}
}

