% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks.r
\name{ndex.get.complete.network}
\alias{ndex.get.complete.network}
\title{Get complete network}
\usage{
ndex.get.complete.network(ndexcon, network_id)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{network_id}{unique ID of the network}
}
\value{
\code{\link{RCX}} object
}
\description{
Get complete network
}
\details{
Uses getEdges (this procedure will return complete network with all elements)
Nodes use primary ID of the base term ('represents' element)
Edges use primary ID of the base term ('predicate', or 'p' element)
Mapping table for the nodes is retrieved ('alias' and 'related' terms) to facilitate conversions/data mapping
Compatible to NDEx server version 1.3 and 2.0
}
\section{REST query}{

This function runs GET query /network/{networkUUID}/asCX  and returns CX network object which is parsed into an RCX object
}
\examples{
\dontrun{
ndexcon = ndex.connect(verbose=T)
pws = ndex.find.networks(ndexcon,"p53")
rcx = ndex.get.complete.network(ndexcon,pws[1,"externalId"]) }
}

