% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks.r
\name{ndex.get.limited.aspect}
\alias{ndex.get.limited.aspect}
\title{Get Some Aspect Elements from a Network as CX}
\usage{
ndex.get.limited.aspect(ndexcon, network_id, aspectName, limit)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{network_id}{unique ID of the network}

\item{aspectName}{string name of the aspect to retrieve}

\item{limit}{integer limiting the amount of output to retrieve}
}
\value{
\code{\link{RCX}} object
}
\description{
This method returns elements in the specified aspect up to the specified limit.
}
\section{REST query}{

This function runs GET query /network/{networkId}/aspect/{aspectName}/{limit}  and returns CX network containing the specified aspect up to the specified limit which is parsed into an RCX object
}
\examples{
\dontrun{
ndexcon = ndex.connect(verbose=T)
pws = ndex.find.networks(ndexcon,"p53")
ndex.get.limited.aspect(ndexcon,pws[1,"externalId"],"citations",5) }
}

