% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_connect.r
\name{ndex.get.network.api}
\alias{ndex.get.network.api}
\title{List possible queries to NDEx API}
\usage{
ndex.get.network.api(ndexcon)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}
}
\value{
data.frame detailing the API names, paths, parameters and athentication needed for API requests.
}
\description{
This function returns a data.frame listing all the possible API calls supported by the NDEx server.
}
\details{
Retrieves information on the NDEx API calls
}
\examples{
\dontrun{
ndexcon = ndex.connect(verbose=T)
ndex.get.network.api(ndexcon)
}
}
\seealso{
\code{\link{ndex_rest_GET}},  \code{\link{ndex_rest_PUT}},  \code{\link{ndex_rest_POST}}
}

