% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfinished_stuff.r
\name{ndex.get.network.namespace}
\alias{ndex.get.network.namespace}
\title{Get the BEL Namespaces Associated with a Network UUID}
\usage{
ndex.get.network.namespace(ndexcon, network_id)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{network_id}{unique ID of the network}
}
\value{
data.frame listing network namespace
}
\description{
This function retrieves the BEL namespaces associated with the supplied network UUID.
}
\section{REST query}{

This function runs GET query /network/{networkId}/namespace    and returns a list of BEL Namespaces
}
\examples{
\dontrun{
ndexcon = ndex.connect(verbose=T)
pws = ndex.find.networks(ndexcon,"p53")
ndex.get.network.namespace(ndexcon,pws[1,"externalId"]) }
}

