% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks.r
\name{ndex.get.network.provenance}
\alias{ndex.get.network.provenance}
\title{Get Network Provenance by Network UUID
GET    /network/{networkUUID}/provenance       Provenance}
\usage{
ndex.get.network.provenance(ndexcon, network_id)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{network_id}{unique ID of the network}
}
\value{
List of network metadata: ID, name, whether it is public, edge and node count; source and format of network
}
\description{
Get Network Provenance by Network UUID
GET    /network/{networkUUID}/provenance       Provenance
}

