% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ndex.get.network.summary}
\alias{ndex.get.network.summary}
\title{Get NetworkSummary by Network UUID
GET    /network/{networkUUID}       NetworkSummary}
\usage{
ndex.get.network.summary(network_id)
}
\arguments{
\item{network_id}{unique ID of the network}
}
\value{
Complete JSON response or Data frame with network metadata: ID, name, whether it is public, edge and node count; source and format of network
}
\description{
Get NetworkSummary by Network UUID
GET    /network/{networkUUID}       NetworkSummary
}

