% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_groups.R
\name{ndex.group.delete.membership}
\alias{ndex.group.delete.membership}
\title{Remove a Group Member}
\usage{
ndex.group.delete.membership(ndexcon, groupId, userId)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{groupId}{character; unique ID (UUID) of the group}

\item{userId}{character; unique ID (UUID) of the user}
}
\value{
Empty string ("") on success, else error
}
\description{
Removes the member from the group
}
\note{
Compatible to NDEx server version 2.0
}
\section{REST query}{

DELETE: ndex.api.config$api$user$membership$delete
}

\examples{
\dontrun{
groupId = "aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee"
userId = "ffffffff-gggg-hhhh-iiii-jjjjjjjjjjjj"
ndex.group.delete.membership(ndexcon, groupId, userId)
}
}
