% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_groups.R
\name{ndex.group.get.network}
\alias{ndex.group.get.network}
\title{Get Group Permission for a Specific Network}
\usage{
ndex.group.get.network(ndexcon, groupId, networkId)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{groupId}{character; unique ID (UUID) of the group}

\item{networkId}{character; unique ID (UUID) of the network}
}
\value{
Network permissions of that group or empty object
}
\description{
Get Group Permission for a Specific Network
}
\note{
Compatible to NDEx server version 2.0
}
\section{REST query}{

GET: ndex.api.config$api$group$network$get
}

\examples{
\dontrun{
groupId = "aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee"
networkId = "aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee"
group = ndex.group.get.network(ndexcon, groupId, networkId)
}
}
