% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_groups.R
\name{ndex.group.set.membership}
\alias{ndex.group.set.membership}
\title{Add or Update a Group Member}
\usage{
ndex.group.set.membership(ndexcon, groupId, userId, type = "MEMBER")
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{groupId}{character; unique ID (UUID) of the group}

\item{userId}{character; unique ID (UUID) of the user}

\item{type}{character (optional)("GROUPADMIN"|"MEMBER")(default: "MEMBER"); Type of group membership}
}
\value{
Empty string ("") on success, else error
}
\description{
Updates the membership corresponding to the GroupMembership type specified in the URL parameter.
}
\note{
Compatible to NDEx server version 2.0
}
\section{REST query}{

PUT: ndex.api.config$api$user$membership$update
}

\examples{
\dontrun{
groupId = "aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee"
userId = "ffffffff-gggg-hhhh-iiii-jjjjjjjjjjjj"
ndex.group.set.membership(ndexcon, groupId, userId)
ndex.group.set.membership(ndexcon, groupId, userId, type='MEMBER')	## same as before
ndex.group.set.membership(ndexcon, groupId, userId, type='GROUPADMIN')
}
}
