% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_helper.r
\name{ndex.helper.encodeParams}
\alias{ndex.helper.encodeParams}
\title{Adds Parameters to an url}
\usage{
ndex.helper.encodeParams(url, params, ...)
}
\arguments{
\item{url}{character}

\item{params}{(named) character vector;}

\item{values}{character vector (optional if params are named);}

\item{encoding}{character (optional); Method to encode the parameter: 'asParams', 'asNamedParams', 'asPath', 'withinURL';}
}
\value{
URL with encoded parameters as character
}
\description{
Encodes the given parameter and adds it to the url. The methods can be 'asParams', 'asPath' and 'withinURL'.
The methods are chosen automatically depending on the provided parameter:
if only values are provided, the values are encoded as url (e.g. "url/value1/value2/...")[encoding='asPath']
if the params are unnamed and values are provided, it is encoded as params (e.g. "url?param1=value1&param2=value2&...")[encoding='asParams']
if the params are named an no values are provided, it is encoded as params (e.g. "url?param1=value1&param2=value2&...")[encoding='asNamedParams']
if named params and values are given, the value of params is used to replace it in the url with the corresponding value (e.g. url="abc.de/#test#/else", params=(bla='#test#') and value=('blubb') becomes to "abc.de/blubb/else")[encoding='withinURL']
}
\note{
params and values must have the same length
}
\examples{
\dontrun{
url = "http://en.wikipedia.org/w/index.php"
values = c("Train", "5", "90", "history")
ndex.helper.encodeParams(url, values=values)
#[1] "http://en.wikipedia.org/w/index.php/Train/5/90/history"

params = c("title", "limit", "offset", "action") 
namedParams = c(title="Train", limit="5", offset="90", action="history")
ndex.helper.encodeParams(url, params=params, values=values)
ndex.helper.encodeParams(url, params, values)          ## same as above, but shorter
ndex.helper.encodeParams(url, params=namedParams)      ## same as above, but with named params
ndex.helper.encodeParams(url, namedParams)             ## same as above, but shorter
#[1] "http://en.wikipedia.org/w/index.php?title=Train&limit=5&offset=90&action=history"

url = "http://en.wikipedia.org/w/index.php/#Train#/somethingElse/#Number#"
namedParams = c(train="#Train#", someNumber="#Number#")
values = c("ICE200", 12345)
ndex.helper.encodeParams(url, params=namedParams, values=values)
ndex.helper.encodeParams(url, namedParams, values)
#[1] "http://en.wikipedia.org/w/index.php/ICE200/somethingElse/12345"
}
}

