% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_helper.r
\name{ndex.helper.getApi}
\alias{ndex.helper.getApi}
\title{Get the Api configuration for a function}
\usage{
ndex.helper.getApi(ndexcon, apiPath)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection \code{\link{ndex.connect}}}

\item{apiPath}{character; paht to follow in the nested list}
}
\value{
configuration of the function
}
\description{
This function extracts the function definition from the ndex configuration within a ndex-connection object.
It follows the given path down the list.
}
\note{
This function is internal.
}
\examples{
## Establish a server connection
ndexcon = ndex.connect()
## Get the function definition for ndex.network.get.summary
## ndex.conf[[ndex.conf$defaultVersion]]$api$network$summary$get
ndex.helper.getApi(ndexcon, 'network$summary$get')
}
