% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_helper.r
\name{ndex.helper.httpResponseHandler}
\alias{ndex.helper.httpResponseHandler}
\title{Handles the http server response}
\usage{
ndex.helper.httpResponseHandler(response, description, verbose = F)
}
\arguments{
\item{response}{object of class response (httr)}

\item{description}{character; description of the action performed}

\item{verbose}{logical; whether to print out extended feedback}
}
\description{
This function handles the response from a server. If some response code different from success (200) is returned, the execution stops and the reason is shown.
}
\examples{
\dontrun{
 ndex.helper.httpResponseHandler(httr::GET('http://www.ndexbio.org'), 'Tried to connect to NDEx server', T)
 }
}

