% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_helper.r
\name{ndex.helper.UrlAddParams}
\alias{ndex.helper.UrlAddParams}
\title{Adds Parameters to an url}
\usage{
ndex.helper.UrlAddParams(url, params, values)
}
\arguments{
\item{url}{string}

\item{params}{character vector;}

\item{values}{character vector;}
}
\value{
URL with parameters as string
}
\description{
This functions searches the public networks on an NDEx server for networks containing the supplied search string. T
his search can be limited to certain accounts as well as in length.
}
\note{
params and values must have the same length
}
\examples{
\dontrun{
ndex.helper.UrlAddParams("http://en.wikipedia.org/w/index.php", c("title", "limit", "offset", "action"), c("Train", "5", "90", "history"))
# "http://en.wikipedia.org/w/index.php?title=Train&limit=5&offset=90&action=history"
}
}

