% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_RCX.r
\name{ndex.JSON2RCX}
\alias{ndex.JSON2RCX}
\title{Create RCX object from JSON data}
\usage{
ndex.JSON2RCX(json, verbose = FALSE)
}
\arguments{
\item{json}{JSON data}

\item{verbose}{logical; whether to print out extended feedback}
}
\value{
returns object of class RCX if successfull, NULL otherwise
}
\description{
Create RCX object from JSON data
}

