% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks_aspects_and_metadata.r
\name{ndex.network.get.aspects}
\alias{ndex.network.get.aspects}
\title{Get Network CX Metadata Collection}
\usage{
ndex.network.get.aspects(ndexcon, network_id)
}
\description{
Get Network CX Metadata Collection
}

