% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks_aspects_and_metadata.r
\name{ndex.network.get.metadata}
\alias{ndex.network.get.metadata}
\title{Get Network CX Metadata Collection}
\usage{
ndex.network.get.metadata(ndexcon, networkId)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{networkId}{character; unique ID (UUID) of the network}
}
\value{
metadata as list: consistencyGroup, elementCount, lastUpdate, name, properties, version and idCounter
}
\description{
This function retrieves the (aspect) meta-data of the network identified by the supplied network UUID string.
}
\note{
Compatible to NDEx server version 1.3 and 2.0, but doesn't work for version 1.3
}
\section{REST query}{

GET: ndex.api.config$api$network$aspect$getMetaData
}

\examples{
\dontrun{
ndexcon = ndex.connect(verbose=T)
pws = ndex.find.networks(ndexcon,"p53")
ndex.network.get.metadata(ndexcon,pws[1,"externalId"]) 
}
}
