% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks_ndex_properties.r
\name{ndex.network.get.samples}
\alias{ndex.network.get.samples}
\title{Get Network Sample}
\usage{
ndex.network.get.samples(ndexcon, networkId)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{networkId}{character; unique ID (UUID) of the network}
}
\value{
Network sample
}
\description{
Get Network Sample
}
