% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks_simple_operations.r
\name{ndex.network.get.summary}
\alias{ndex.network.get.summary}
\title{Get NetworkSummary by Network UUID}
\usage{
ndex.network.get.summary(ndexcon, nuuid)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{nuuid}{unique ID of the network}
}
\value{
List of network metadata: ID, name, whether it is public, edge and node count; source and format of network
}
\description{
This function retrieves the summary of the network identified by the supplied network UUID string.
}
\note{
Compatible to NDEx server version 1.3 and 2.0
}
\section{REST query}{

GET: ndex.api.config$api$network$summary$get
}

\examples{
\dontrun{
ndexcon = ndex.connect(verbose=T)
pws = ndex.find.networks(ndexcon,"p53")
ndex.network.get.summary(ndexcon,pws[1,"externalId"]) }
}
