% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks_ndex_properties.r
\name{ndex.network.set.properties}
\alias{ndex.network.set.properties}
\title{Set Network Properties}
\usage{
ndex.network.set.properties(ndexcon, nuuid)
}
\description{
Set Network Properties
}
