% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks_ndex_properties.r
\name{ndex.network.set.provenance}
\alias{ndex.network.set.provenance}
\title{Set Network Provenance}
\usage{
ndex.network.set.provenance(ndexcon, network_id)
}
\description{
Set Network Provenance
}

