% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks_ndex_properties.r
\name{ndex.network.set.samples}
\alias{ndex.network.set.samples}
\title{Set Sample Network}
\usage{
ndex.network.set.samples(ndexcon, networkId)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{networkId}{character; unique ID (UUID) of the network}
}
\value{
NULL on success
}
\description{
Set Sample Network
}
