% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks_aspects_and_metadata.r
\name{ndex.network.update.aspect}
\alias{ndex.network.update.aspect}
\title{Update an Aspect of a Network}
\usage{
ndex.network.update.aspect(ndexcon, networkId, aspectName, aspectAsRCX)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection link{ndex.connect}}

\item{networkId}{unique ID of the network}

\item{aspectName}{name of the aspect}

\item{aspectAsRCX}{rcx data for the aspect (rcx[[aspectName]])}
}
\value{
networkId unique ID of the modified network
}
\description{
This function updates an aspect with the provided CX for the aspect.
}
\note{
Requires an authorized user! (ndex.connect with credentials)

Compatible to NDEx server version 2.0
}
\section{REST query}{

PUT: ndex.conf$api$network$aspect$update
}

\examples{
## Establish a server connection with credentials 
# ndexcon = ndex.connect('MyAccountName', 'MyPassword')
## Find one of your networks and get its UUID
# networks = ndex.find.networks(ndexcon, accountName='MyAccountName')
# networkId = networks[1,"externalId"]
## Get the network data 
# aspect = ndex.network.get.aspect(ndexcon, networkId, 'nodeAttributes') 
## Do some changes to the aspect..
# aspectModified = aspect[1:5,]
## and update the aspect
# ndex.network.update.aspect(ndexcon,pws[1,"externalId"], 'nodeAttributes', aspectModified)
NULL
}
