% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks_ndex_properties.r
\name{ndex.network.update.permission}
\alias{ndex.network.update.permission}
\title{Update Network Permission}
\usage{
ndex.network.update.permission(ndexcon, network_id)
}
\description{
Update Network Permission
}

