% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_cbdd.r
\name{ndex.ngraph2CBDD}
\alias{ndex.ngraph2CBDD}
\title{Create CBDD-like data from \code{\link{ngraph}} object}
\usage{
ndex.ngraph2CBDD(ngraph, useNamespace = NULL)
}
\arguments{
\item{ngraph}{\code{\link{ngraph}} object}

\item{useNamespace}{string (optional) name of the namespace for nodes}
}
\value{
data.frame with two columns containing IDs of interacting nodes. Returns NULL if ngraph object does not contain edge information.
}
\description{
Returns CBDD-like data as defined at https://cbdd.thomsonreuterslifesciences.com/cbdd/help_algorithms/
The identifiers used are (in this order, if applicable) the namespace ids, the "r"epresents field, the "n"ame field or the @id field of the CX data
}
\details{
Internally this function calls ndex.ngraph2RCX and then ndex.RCX2CBDD.
}
\examples{
\dontrun{
ndexcon = ndex.connect(verbose=T)
pws = ndex.find.networks(ndexcon,"p53")
rcx = ndex.get.complete.network(ndexcon,pws[1,"externalId"]) 
ngraph = ndex.RCX2ngraph(rcx) 
cbdd = ndex.ngraph2CBDD(ngraph)}
}

