% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_ngraph.r
\name{ndex.ngraph2RCX}
\alias{ndex.ngraph2RCX}
\title{Create RCX object from ngraph object}
\usage{
ndex.ngraph2RCX(ngraph, verbose = FALSE)
}
\arguments{
\item{ngraph}{ngraph object}

\item{verbose}{logical; whether to print out extended feedback}
}
\value{
returns object of class RCX if successfull, NULL otherwise
}
\description{
Create RCX object from ngraph object
}

