% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ndex.property.graph.as.ndexgraph}
\alias{ndex.property.graph.as.ndexgraph}
\title{Convert propertyGraph export to ndexgraph}
\usage{
ndex.property.graph.as.ndexgraph(property_graph)
}
\arguments{
\item{property_graph}{Property graph data structure (as returned by relevant API functions)}
}
\value{
\code{ndexgraph} object
}
\description{
Convert propertyGraph export to ndexgraph
}
\details{
Collects term ID, name and properties for nodes;
predicate, subject, term ID and properties for edges.
Each property takes up a column;
column name is predicateString
and value is \code{value} for given node/edge
}
\note{
PropertyGraphNetwork JSON data are more easily converted to ndexgraph objects than Network JSON data
}
\examples{
\dontrun{
x <- ndex.get.complete.network.as.property.graph('32733ca9-4e84-11e4-98fb-000c29873918')
y <- ndex.property.graph.as.ndexgraph(x)
}
}

