% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_ngraph.r
\name{ndex.RCX2ngraph}
\alias{ndex.RCX2ngraph}
\title{Create ngraph object from RCX object}
\usage{
ndex.RCX2ngraph(rcx, verbose = FALSE)
}
\arguments{
\item{rcx}{RCX object}

\item{verbose}{logical; whether to print out extended feedback}
}
\value{
returns object of class ngraph if successfull, NULL otherwise
}
\description{
Create ngraph object from RCX object
}

