% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_RCX.r
\name{ndex.RCXasNewNetwork}
\alias{ndex.RCXasNewNetwork}
\title{Remove all interfering NDEx artefacts from RCX object}
\usage{
ndex.RCXasNewNetwork(rcx)
}
\arguments{
\item{rcx}{RCX object}
}
\value{
\code{\link{RCX}} object
}
\description{
Remove all interfering NDEx artefacts from RCX object
}
\details{
After a RCX is downloaded from an NDEx server, it will contain some aspects that are not present in a newly generated network, i.e. ndexStatus', provenanceHistory' and 'status'.
Removing those aspects might be usefull in some cases.
}
\examples{
\dontrun{
ndexcon = ndex.connect(verbose=T)
pws = ndex.find.networks(ndexcon,"p53")
rcx = ndex.get.complete.network(ndexcon,pws[1,"externalId"]) 
rcx = ndex.RCXAsNewNetwork(rcx)
rcxjson = ndex.RCX2JSON(rcx)
ndex.create.network(ndexcon, rcxjson) }
}
\seealso{
\code{\link{ndex.JSON2RCX}} \code{\link{ndex.get.complete.network}}
}

