% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_RCX.r
\name{ndex.RCXmerge}
\alias{ndex.RCXmerge}
\title{merging two or more rcx objects
||__layerMapping=NULL__|Layers are just megred toghether. If two rcx objects contain thesame layer id (by accident), the two layers are combined to one.
__layerMapping=c('a','b')__|The vector elements are used as prefix for the original ids,i.e. with two input rcx objects and layers '1', '2' and '3' each,the layers of the first object become 'a1', 'a2' and 'a3', and forthe second object 'b1', 'b2' and 'b3' respectively
__layerMapping=function(rcxIndex, layerId)__|Names the layers using a function, which is dependent on theindex of rcx object in the mergeRCX function call, and the layerid within that object. If two return values of the mappingfunction are equal, the two layers are combined to one.||}
\usage{
ndex.RCXmerge(rcx, ...)
}
\description{
merging two or more rcx objects
||__layerMapping=NULL__|Layers are just megred toghether. If two rcx objects contain thesame layer id (by accident), the two layers are combined to one.
__layerMapping=c('a','b')__|The vector elements are used as prefix for the original ids,i.e. with two input rcx objects and layers '1', '2' and '3' each,the layers of the first object become 'a1', 'a2' and 'a3', and forthe second object 'b1', 'b2' and 'b3' respectively
__layerMapping=function(rcxIndex, layerId)__|Names the layers using a function, which is dependent on theindex of rcx object in the mergeRCX function call, and the layerid within that object. If two return values of the mappingfunction are equal, the two layers are combined to one.||
}

