% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_RCX.r
\name{ndex.RCXnextEdgeId}
\alias{ndex.RCXnextEdgeId}
\title{Get the next id, that should be used for creating a new edge
(euqals the maximal id in nodes + 1)}
\usage{
ndex.RCXnextEdgeId(rcx)
}
\arguments{
\item{rcx}{RCX object}
}
\value{
integer
}
\description{
Get the next id, that should be used for creating a new edge
(euqals the maximal id in nodes + 1)
}
\details{
Get the next id, that should be used for creating a new edge
(euqals the maximal id in nodes + 1)
}
\examples{
\dontrun{
ndex.RCXnextEdgeId(rcx) }
}

