% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfinished_stuff.r
\name{ndex.RCXnextNodeId}
\alias{ndex.RCXnextNodeId}
\title{Get the next id, that should be used for creating a new node
(euqals the maximal id in nodes + 1)}
\usage{
ndex.RCXnextNodeId(rcx)
}
\arguments{
\item{rcx}{RCX object}
}
\value{
integer
}
\description{
Get the next id, that should be used for creating a new node
(euqals the maximal id in nodes + 1)
}
\details{
Get the next id, that should be used for creating a new node
(euqals the maximal id in nodes + 1)
}
\examples{
\dontrun{
ndex.RCXnextNodeId(rcx) }
}

