% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfinished_stuff.r
\name{ndex.set.network.readonly}
\alias{ndex.set.network.readonly}
\title{Set a Network Read-Only}
\usage{
ndex.set.network.readonly(ndexcon, network_id, readonly = c(TRUE, FALSE)[1])
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{network_id}{unique ID of the network}

\item{readonly}{logical defaults to TRUE}
}
\value{
response from server
}
\description{
This function sets the readonly flag on the network identified by the supplied network UUID.
}
\section{REST query}{

This function runs GET query /network/{networkId}/setFlag/{parameter}={value}
}
\examples{
\dontrun{
ndexcon = ndex.connect(verbose=T)
pws = ndex.find.networks(ndexcon,"p53")
ndex.set.network.readonly(ndexcon,pws[1,"externalId"],readonly=TRUE) }
}

