% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_groups.R
\name{ndex.update.group}
\alias{ndex.update.group}
\title{Update Group}
\usage{
ndex.update.group(ndexcon, groupId, groupName, image, website, description,
  properties)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{groupId}{character; unique ID (UUID) of the group}

\item{groupName}{character; name of the new graoup}

\item{image}{character (optional); URL of the account owner's image.}

\item{website}{character (optional); URL of the account owner's web site}

\item{description}{character (optional); Short description of this user.}

\item{properties}{list (optional); additional properties for the group}
}
\value{
Empty string ("") on success, else error
}
\description{
Updates the group based on the data.
}
\note{
Compatible to NDEx server version 2.0
}
\section{REST query}{

PUT: ndex.api.config$api$user$update
}

\examples{
\dontrun{
groupId = "aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee"
ndex.update.group(ndexcon, groupId, description='A really nice group!')
}
}
