% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks.r
\name{ndex.update.network}
\alias{ndex.update.network}
\title{Update an Entire Network as CX}
\usage{
ndex.update.network(ndexcon, rcx, network_id = NULL)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{rcx}{\code{\link{RCX}} object}
}
\value{
\code{\link{RCX}} object
}
\description{
PUT : /network/asCX
This method creates a new network on the NDEx server.
}
\section{REST query}{

POST : /network/asCX
}

