% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_users.R
\name{ndex.user.change.password}
\alias{ndex.user.change.password}
\title{Change Password}
\usage{
ndex.user.change.password(ndexcon, userId, password)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection link{ndex.connect}}

\item{userId}{character; unique ID of the user}

\item{password}{character; New password}
}
\value{
Empty string on success, else error
}
\description{
Changes the authenticated user's password to the new password
}
\note{
Requires an authorized user! (ndex.connect with credentials)

Compatible to NDEx server version 2.0
}
\section{REST query}{

GET: ndex.conf$api$user$password$change
}

\examples{
## Establish a server connection with credentials 
# ndexcon = ndex.connect('MyAccountName', 'MyPassword')
## Find user and get its id
# user = ndex.find.user.byName(ndexcon, 'SomeUserName')
# userId = user$externalId
## Change user password
# ndex.user.change.password(ndexcon, userId, 'SuperSaveNewPassword')
NULL
}
