% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_users.R
\name{ndex.user.get.networksummary}
\alias{ndex.user.get.networksummary}
\title{Get User's Account Page Networks}
\usage{
ndex.user.get.networksummary(ndexcon, userId)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{userId}{character; unique ID (UUID) of the user}
}
\value{
data.frame of networks (name, description, externalId, uri, etc.) on the account page of the specified user
}
\description{
This is a convenience function designed to support "My Account" pages in NDEx applications. 
It returns a list of NetworkSummary objects to display.
}
\note{
Compatible to NDEx server version 2.0
}
\section{REST query}{

GET: ndex.api.config$api$user$networksummary
}

\examples{
\dontrun{
## get user by name to get UUID
user = ndex.find.user.byName(ndexcon, 'some.ndex.account')

networkSummary = ndex.user.get.networksummary(con, user$externalId)
names(networkSummary)
## [1] "ownerUUID"        "isReadOnly"       "subnetworkIds"    "errorMessage"     "isValid"         
## [6] "warnings"         "isShowcase"       "visibility"       "edgeCount"        "nodeCount"       
##[11] "uri"              "version"          "owner"            "name"             "properties"      
##[16] "description"      "externalId"       "isDeleted"        "modificationTime" "creationTime"
}  
}
