% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_users.R
\name{ndex.user.show.group}
\alias{ndex.user.show.group}
\title{Get User's Membership in Group}
\usage{
ndex.user.show.group(ndexcon, userId, groupId)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection link{ndex.connect}}

\item{userId}{character; unique ID (UUID) of the user}

\item{groupId}{character; unique ID (UUID) of the group}
}
\value{
List of permissions of that user or empty object
}
\description{
Returns the permission that the user specified in the URL has on the given group. Returns an empty object if the authenticated user is not a member of this group.
}
\note{
Requires an authorized user! (ndex.connect with credentials)

Compatible to NDEx server version 2.0
}
\section{REST query}{

GET: ndex.conf$api$user$group$get
}

\examples{
## Establish a server connection with credentials 
# ndexcon = ndex.connect('MyAccountName', 'MyPassword')
## Find user and get its id
# user = ndex.find.user.byName(ndexcon, 'SomeUserName')
# userId = user$externalId
## Find the user's groups and get one group id
# groups = ndex.user.list.groups(ndexcon, userId)
# groupId = groups[1,"externalId"]
## get users's permission in the group
# userPermissions = ndex.user.show.group(ndexcon, userId, groupId)
## $`uuuuuuuu-ssss-eeee-rrrr-123456789abc`
## [1] "MEMBER"
NULL
}
