% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_users.R
\name{ndex.verify.user}
\alias{ndex.verify.user}
\title{Verify a User}
\usage{
ndex.verify.user(ndexcon, userId, code)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection link{ndex.connect}}

\item{userId}{character; unique ID of the user}

\item{code}{character; Verification code sent by email}
}
\value{
string "User account XXX has been activated." when this user's account is successfully activated.
}
\description{
Verify the given user with UUID and verification code, which is set by email
}
\note{
Compatible to NDEx server version 2.0
}
\section{REST query}{

GET: ndex.conf$api$user$verify
}

\examples{
## Establish a server connection
# ndexcon = ndex.connect()
## Find user and get its id
# user = ndex.find.user.byName(ndexcon, 'SomeUserName')
# userId = user$externalId
## Verify user with verification code
# ndex.verify.user(ndexcon, userId, 'Osqy11mRZ9')
## [1] "User account XXX has been activated."
NULL
}
