% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_maintenance.R
\docType{data}
\name{ndex_conf_header}
\alias{ndex_conf_header}
\title{Default header for the ndex_api_config.r file}
\format{An object of class \code{character} of length 1.}
\usage{
ndex_conf_header
}
\value{
character containing the header
}
\description{
Default header for the ndex_api_config.r file
}
\note{
only for package maintenance!
}
\examples{
NULL
}
\keyword{datasets}
