% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_api_config.r
\docType{data}
\name{ndex_config}
\alias{ndex_config}
\title{NDEx server api configuration}
\format{
An object of class \code{list} of length 4.
}
\usage{
ndex_config
}
\value{
Nested list resembling the NDEx server REST API structure
}
\description{
This nested list contains the url and methods for accessing the NDEx server via its REST full api.
It contains specifications for NDEx server api version 1.3 and 2.0. The default api is specified by 'defaultVersion'.
If possible, the version 2.0 should be used.
Own configurations must contain a 'version' entry!
}
\examples{
names(ndex_config$Version_2.0) 
}
\keyword{datasets}
