% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_helper.r
\name{ndex_helper_getApi}
\alias{ndex_helper_getApi}
\title{Get the Api configuration for a function}
\usage{
ndex_helper_getApi(ndexcon, apiPath)
}
\arguments{
\item{ndexcon}{object of class NDExConnection \code{\link{ndex_connect}}}

\item{apiPath}{character; path to follow in the nested list}
}
\value{
configuration of the function
}
\description{
This function extracts the function definition from the ndex configuration within a ndex-connection object.
It follows the given path down the list.
}
\note{
This function is internal.
}
\examples{
## Establish a server connection
ndexcon = ndex_connect()
## Get the function definition for ndex_network_get_summary
## ndex_config[[ndex_config$defaultVersion]]$api$network$summary$get
ndexr:::ndex_helper_getApi(ndexcon, 'network$summary$get')
}
\keyword{internal}
