% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ndexgraph-class}
\alias{ndexgraph-class}
\title{S4 class representing NDEx network}
\description{
S4 class representing NDEx network
}
\section{Slots}{

\describe{
\item{\code{nodes}}{Data frame with node properties}

\item{\code{edges}}{Data frame with edge properties}

\item{\code{properties}}{Data frame containing network properties}

\item{\code{name}}{Name of the network}

\item{\code{id}}{Unique ID of the network}
}}
\note{
So far ID and name are optional (by default will be NA upon initialization, and go unchecked by validator).
}

