% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ndexgraph.as.ndex.property.graph}
\alias{ndexgraph.as.ndex.property.graph}
\title{Convert ndexgraph to propertyGraph for import to NDEx}
\usage{
ndexgraph.as.ndex.property.graph(ndexgraph)
}
\arguments{
\item{ndexgraph}{Object of class \code{\link{ndexgraph}}}
}
\value{
List corresponding to propertyGraph JSON structure
}
\description{
Convert ndexgraph to propertyGraph for import to NDEx
}
\details{
Creates node elements from term ID, name; adds properties if there are any;
Creates edge elements from predicate, subject, object ID, term ID and properties if there are any.
}
\examples{
\dontrun{
x <- ndex.get.complete.network.as.property.graph('32733ca9-4e84-11e4-98fb-000c29873918')
y <- ndex.property.graph.as.ndexgraph(x)
z <- ndexgraph.as.ndex.property.graph(y)
}
}

