% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_ngraph.r
\name{ngraph.toRCX}
\alias{ngraph.toRCX}
\title{Create RCX object from ngraph object}
\usage{
ngraph.toRCX(ngraph, verbose = FALSE)
}
\arguments{
\item{ngraph}{ngraph object}

\item{verbose}{logical; whether to print out extended feedback}
}
\value{
returns object of class RCX if successfull, NULL otherwise
}
\description{
This function creates an RCX object from a valid ngraph object.
}
\details{
The following rules apply to convert from \code{\link{ngraph}} to \code{\link{RCX}}:
\itemize{
 \item all graph attributes are stored as named data.frames within the RCX object
 \item nodes receive their name value as "@id" attribute. All other node attributes are saved in the RCX object as nodeAttributes, access via rcx[["nodeAttributes"]]. 
       Data goes from wide format (columns for each unique n with cells contianing v) to long format (column n containing attribute name and column v containing attribute value).
 \item edges are connected via their "s"art and "t"arget fields. The "@id" and "i"nteraction attribute are stored as is and all edgeAttributes are saved as node attributes, access via rcx[["edgeAttributes"]]. 
       Data goes from wide format (columns for each unique n with cells contianing v) to long format (column n containing attribute name and column v containing attribute value).
}
}
\examples{
## Establish a server connection with credentials 
# ndexcon = ndex.connect('MyAccountName', 'MyPassword')
## Find one of your networks and get its UUID
# networks = ndex.find.networks(ndexcon, accountName='MyAccountName')
# networkId = networks[1,"externalId"]
## Get the network data 
rcx = ndex.get.network(ndexcon, networkId) 
## Convert to nGraph
ngraph = rcx.toNGraph(rcx) 
## Convert it back to rcx
rcx = ngraph.toRCX(ngraph)
}
\seealso{
\code{\link{ngraph}} \code{\link{ngraph.fromRCX}} \code{\link{rcx.fromJSON}} \code{\link{rcx.toJSON}}
}
