% Generated by roxygen2 (4.0.2): do not edit by hand
\name{properties2df}
\alias{properties2df}
\title{Convert properties of nodes / edges to a data frame}
\usage{
properties2df(entities, missing.value = NA)
}
\arguments{
\item{entities}{List of node or edge elements from 'propertyGraph'}

\item{missing.value}{what to insert when a property is missing for a given element; default NA}
}
\value{
Data frame with all properties; missing values defined by missing.value
}
\description{
Convert properties of nodes / edges to a data frame
}
\details{
Each node or edge can have properties (possibly not the same number of values across nodes/edges)
Our goal is to turn them to the data frame, covering all properties and putting NA for the missing properties if those happen.
}
\examples{
\dontrun{
}
}

