% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_RCX.r
\name{rcx.asNewNetwork}
\alias{rcx.asNewNetwork}
\title{Remove all interfering NDEx artefacts from RCX object}
\usage{
rcx.asNewNetwork(rcx)
}
\arguments{
\item{rcx}{RCX object}
}
\value{
\code{\link{RCX}} object
}
\description{
Remove all interfering NDEx artefacts from RCX object
}
\details{
After a RCX object is downloaded from an NDEx server, it will contain some aspects that are not present in a newly generated network, i.e. ndexStatus', provenanceHistory' and 'status'.
Removing those aspects might be useful in some cases.
}
\examples{
\dontrun{
ndexcon = ndex.connect(verbose=T)
pws = ndex.find.networks(ndexcon,"p53")
rcx = ndex.get.network(ndexcon,pws[1,"externalId"]) 
rcx = rcx.asNewNetwork(rcx)
rcxjson = rcx.toJSON(rcx)
ndex.create.network(ndexcon, rcxjson) }
}
\seealso{
\code{\link{rcx.fromJSON}} \code{\link{ndex.get.network}}
}
