% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_RCX.r
\name{rcx.merge}
\alias{rcx.merge}
\title{merging two or more rcx objects}
\usage{
rcx.merge(rcx, ...)
}
\arguments{
\item{rcx}{RCX object}

\item{...}{one or more RCX objects}
}
\value{
RCX object
}
\description{
merging two or more rcx objects
}
\details{
Two or more RCX objects are merged
}

