% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_RCX.r
\name{rcx.new}
\alias{rcx.new}
\title{Create a blank rcx object}
\usage{
rcx.new(nodes)
}
\arguments{
\item{nodes}{At least one node is necessary for be a valid rcx}
}
\value{
RCX object
}
\description{
Create a blank rcx object
}
\details{
Constructor for creating a blank rcx object, that fulfills the minimal requirements of CX
}

