% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_RCX.r
\name{rcx.toJSON}
\alias{rcx.toJSON}
\title{Generate JSON data from RCX object}
\usage{
rcx.toJSON(rcx, verbose = FALSE, pretty = FALSE)
}
\arguments{
\item{rcx}{RCX object}

\item{verbose}{logical; whether to print out extended feedback}

\item{pretty}{logical; adds indentation whitespace to JSON output}
}
\value{
json jsonlite json object if successfull, NULL otherwise
}
\description{
Generate JSON data from RCX object
}
\examples{
## Establish a server connection
ndexcon = ndex.connect()
## Find a network and get its UUID
networks = ndex.find.networks(ndexcon,"p53")
networkId = networks[1,"externalId"]
## Get the network data 
rcx = ndex.get.network(ndexcon, networkId)
## Convert RCX to JSON
rcxjson = rcx.toJSON(rcx)
}
\seealso{
\code{\link{ngraph.fromRCX}} \code{\link{ngraph.toRCX}} \code{\link{rcx.fromJSON}}
}
