% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_RCX.r
\name{rcx.updateMetaData}
\alias{rcx.updateMetaData}
\title{Updating the meta-data of an RCX object}
\usage{
rcx.updateMetaData(rcx, mandatoryAspects = c("nodes"),
  excludeAspects = c("metaData", "numberVerification", "status"),
  force = FALSE, verbose = FALSE)
}
\arguments{
\item{rcx}{RCX object}

\item{mandatoryAspects}{character vector; Aspects, that are mandatory for a valid RCX object (by default: "nodes")}

\item{excludeAspects}{character vector; Aspects, that are excluded for generating metaData (by default: "metaData", "numberVerification" and "status")}

\item{force}{logical; force the creation of new metaData (even if the RCX object already contains metaData)}

\item{verbose}{logical; whether to print out extended feedback}
}
\value{
\code{\link{RCX}} object
}
\description{
Updating the meta-data of an RCX object
}
\details{
For a given RCX object the meta-data is updated, i.e. the counted elements and id counter are updated. If an aspect was added/removed, it will also added/removed from the meta-data.
If mandatory aspects (specified in mandatoryAspects parameter) are missing in the RCX object, an error is thrown.
}
\examples{
rcx = rcx.updateMetaData(rcx)
# or with explicitly set default values
rcx = rcx.updateMetaData(rcx, mandatoryAspects=c('nodes'), excludeAspects=c("metaData", "numberVerification", "status"), force=FALSE, verbose=FALSE)
}
