% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_RCX.r
\name{rcx_aspect_toJSON}
\alias{rcx_aspect_toJSON}
\title{Generate JSON data for a single aspect of a RCX object}
\usage{
rcx_aspect_toJSON(rcxAspect, verbose = FALSE, pretty = FALSE)
}
\arguments{
\item{rcxAspect}{aspect in RCX object (rcx[[aspectName]])}

\item{verbose}{logical; whether to print out extended feedback}

\item{pretty}{logical; adds indentation whitespace to JSON output}
}
\value{
json object if successfull, empty string otherwise
}
\description{
Generate JSON data for a single aspect of a RCX object
}
\examples{
## Establish a server connection
ndexcon = ndex_connect()
## Find a network and get its UUID
networks = ndex_find_networks(ndexcon,"p53")
networkId = networks[1,"externalId"]
## Get the network data 
rcx = ndex_get_network(ndexcon, networkId)
## Convert RCX aspect to JSON
rcxNodesJson = ndexr:::rcx_aspect_toJSON(rcx$nodes)
}
\seealso{
\code{\link{rcx_toJSON}} and \code{\link{rcx_fromJSON}}
}
