% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_ngraph.r
\name{rcx_toNGraph}
\alias{rcx_toNGraph}
\title{Create ngraph object from RCX object}
\usage{
rcx_toNGraph(rcx, verbose = FALSE)
}
\arguments{
\item{rcx}{RCX object}

\item{verbose}{logical; whether to print out extended feedback}
}
\value{
returns object of class ngraph if successfull, NULL otherwise
}
\description{
Create ngraph object from RCX object
}
\note{
Wrapper function for \code{\link{ngraph_fromRCX}}
}
\examples{
## Establish a server connection
ndexcon = ndex_connect()
## Find one of your networks and get its UUID
networks = ndex_find_networks(ndexcon)
networkId = networks[1,"externalId"]
## Get the network data 
rcx = ndex_get_network(ndexcon, networkId) 
## Convert to nGraph
ngraph = rcx_toNGraph(rcx) 
}
