% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_RCXgraph.r
\name{rcxgraph_toRCX}
\alias{rcxgraph_toRCX}
\title{Create RCX object from RCXgraph object}
\usage{
rcxgraph_toRCX(rcxgraph, verbose = FALSE)
}
\arguments{
\item{rcxgraph}{RCXgraph object}

\item{verbose}{logical; whether to print out extended feedback}
}
\value{
returns object of class RCX if successfull, NULL otherwise
}
\description{
This function creates an RCX object from a valid RCXgraph object.
}
\details{
The following rules apply to convert from \code{\link{RCXgraph}} to \code{\link{RCX}}:
\itemize{
 \item all graph attributes are stored as named data.frames within the RCX object
 \item nodes receive their name value as "@id" attribute. All other node attributes are saved in the RCX object as nodeAttributes, access via rcx[["nodeAttributes"]].
       Data goes from wide format (columns for each unique n with cells contianing v) to long format (column n containing attribute name and column v containing attribute value).
 \item edges are connected via their "s"art and "t"arget fields. The "@id" and "i"nteraction attribute are stored as is and all edgeAttributes are saved as node attributes, access via rcx[["edgeAttributes"]].
       Data goes from wide format (columns for each unique n with cells contianing v) to long format (column n containing attribute name and column v containing attribute value).
}
}
\examples{
## Create an RCX object
rcx = rcx_new(c('@id'=1, n='Some Name', r='HGNC:Symbol'))
## Convert to RCXgraph
rcxgraph = rcxgraph_fromRCX(rcx)
## Convert it back to rcx
rcx = rcxgraph_toRCX(rcxgraph)
}
\seealso{
\code{\link{RCXgraph}} \code{\link{rcxgraph_fromRCX}} \code{\link{rcx_fromJSON}} \code{\link{rcx_toJSON}}
}
