% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_maintenance.R
\name{readYamlToRFile}
\alias{readYamlToRFile}
\title{Translates a YAML file to a R script
Run this manually if you want to update the ndex.api.conf
only for package maintenance!}
\usage{
readYamlToRFile(yamlFile, rFile, fileHeader = "")
}
\arguments{
\item{yamlFile}{character; input file in YAML format}

\item{rFile}{character; output file for the R script}

\item{fileHeader}{character (optional); text that will be put in front of the R script}
}
\description{
Translates a YAML file to a R script
Run this manually if you want to update the ndex.api.conf
only for package maintenance!
}
\examples{
\dontrun{readYamlToRFile('R/ndex_api_config.yml', 'R/ndex_api_config.r', ndex.api.config.header)}
}

