% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_maintenance.R
\name{updateConfigFromYaml}
\alias{updateConfigFromYaml}
\title{Translates a YAML file to a R script
Run this manually if you want to update the ndex.api.conf
only for package maintenance!}
\usage{
updateConfigFromYaml(yamlFile = "R/ndex_api_config.yml",
  rScriptFile = "R/ndex_api_config.r",
  defaultHeader = ndex.api.config.header)
}
\arguments{
\item{yamlFile}{character; input file in YAML format}

\item{rScriptFile}{character; output file for the R script}

\item{defaultHeader}{character (optional); text that will be put in front of the R script}
}
\description{
Translates a YAML file to a R script
Run this manually if you want to update the ndex.api.conf
only for package maintenance!
}
\examples{
\dontrun{updateConfigFromYaml('R/ndex_api_config.yml', 'R/ndex_api_config.r', ndex.api.config.header)}
}

