% Generated by roxygen2 (4.0.2): do not edit by hand
\name{aux_list}
\alias{aux_list}
\alias{jsonlist2df}
\alias{jsonlist2vector}
\alias{unlist2df}
\title{Misc operations with lists (helps converting JSON to tabular formats)}
\usage{
jsonlist2vector(l)

jsonlist2df(l)

unlist2df(l, names = c("key", "value"))
}
\arguments{
\item{l}{list}

\item{names}{character vector of length 2 with column names for output}
}
\description{
Misc operations with lists (helps converting JSON to tabular formats)
}
\details{
\code{jsonlist2df} flattens JSON-derived list and returns a vector. NULLs are replaced with emply string (""); non-vector elements and vector elements with length > 1 are dropped

\code{jsonlist2df} converts JSON-derived lists of the same structure into a data frame. each row corresponds to first order element in l
 for each row, NULLs are replaced with emply string (""); non-vector elements and vector elements with length > 1 are dropped

\code{unlist2df} turns named list into a data frame where 1st column has names of list elements and 2nd has corrsponding values. List should be named and should contain vectors. Non-vector elements will be dropped
}
\examples{
jsonlist2vector(list(a=1, b=NULL, c=1:3, d='v'))
}

