% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_connect.r
\name{ndex.alive}
\alias{ndex.alive}
\title{Check if user is authenticated to NDEx REST server}
\usage{
ndex.alive(ndexcon)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}
}
\value{
logical (TRUE if user is authenticated and connection is active, FALSE otherwise)
}
\description{
This function checks if the supplied NDEXConnection object allows user access to the NDEx server. It will fail for anonymous NDEXConnection objects.
}
\examples{
\dontrun{
 ndexcon = ndex.connect(verbose=T)
 ndex.alive(ndexcon) # should return FALSE since ndexcon is anonymous
 }
}

