% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ndex.alive}
\alias{ndex.alive}
\title{Check if user is authenticated to NDEx REST server}
\usage{
ndex.alive()
}
\value{
logical (TRUE if user is authenticated and connection is active, FALSE otherwise)
}
\description{
Check if user is authenticated to NDEx REST server
}

