% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_connect.r
\name{ndex.connect}
\alias{ndex.connect}
\title{Connect to NDEx REST API}
\usage{
ndex.connect(username, password, host = "http://www.ndexbio.org/rest",
  verbose = FALSE)
}
\arguments{
\item{username}{character username}

\item{password}{character password}

\item{host}{(optional) URL of NDEx REST server to be used; Default is: http://www.ndexbio.org/rest}

\item{verbose}{logical; whether to print out extended feedback}
}
\value{
returns object of class NDEXConnection which stores options and authentication if successfull, NULL otherwise
}
\description{
Connect to NDEx REST API
}

