% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_connect.r
\name{ndex.connect}
\alias{ndex.connect}
\title{Connect to NDEx REST API}
\usage{
ndex.connect(username, password,
  host = "ndex.api.conf$defaults$connection$host",
  apiPath = "ndex.api.conf$defaults$connection$api",
  apiConfig = ndex.api.conf$Version_2.0, verbose = T)
}
\arguments{
\item{username}{character username}

\item{password}{character password}

\item{host}{character (optional) URL of NDEx REST server; Set in ndex.conf: By default: ndex.conf$url$host + ndex.conf$url$path$<api-version>; For version 2.0 the host is "http://www.ndexbio.org/v2", for version 1.3 it is "http://www.ndexbio.org/rest"}

\item{verbose}{logical; whether to print out extended feedback}

\item{apiversion}{character (optional) Version of NDEx REST server; Set in ndex.conf:}
}
\value{
returns object of class NDEXConnection which stores options and authentication if successfull, NULL otherwise
}
\description{
This function creates an NDEXConnection which stores options and authentication details. It is a parameter required for most of the other ndexr functions.
If username and password are missing an anonymous connection is created, which already offers most of the retrieval functionality.
}
\examples{
\dontrun{ndexcon = ndex.connect(verbose=T)}
ToDo: check the ndexcon properties, especially curlOpts
}

