% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ndex.connect}
\alias{ndex.connect}
\title{Connect to NDEx REST API}
\usage{
ndex.connect(username, password, host)
}
\arguments{
\item{username}{character username}

\item{password}{character password}

\item{host}{(optional) URL of NDEx REST server to be used}
}
\value{
returns nothing; RCurl options object for authentication is stored in the special environment and reused with the queries
}
\description{
Connect to NDEx REST API
}
\note{
REST server location can also be set separately using \code{\link{ndex.get.host}}. In this case, supplying host here is not necessary
}
\seealso{
\code{\link{ndex.get.host}}
}

