% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks_simple_operations.r
\name{ndex.create.network}
\alias{ndex.create.network}
\title{Create a Network from CX data}
\usage{
ndex.create.network(ndexcon, rcx)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{rcx}{\code{\link{RCX}} object}
}
\value{
\code{\link{RCX}} object
}
\description{
This method creates a new network on the NDEx server.
Compatible to NDEx server version 1.3 and 2.0
}
\section{REST query}{

POST : /network/asCX
}

