% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks_simple_operations.r
\name{ndex.create.network}
\alias{ndex.create.network}
\title{Create a Network at a server from RCX data}
\usage{
ndex.create.network(ndexcon, rcx)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{rcx}{\code{\link{RCX}} object}
}
\value{
UUID of the newly created network
}
\description{
Create a Network at a server from RCX data
}
\details{
This method creates a new network on the NDEx server from the given RCX object
}
\note{
Compatible to NDEx server version 1.3 and 2.0
}
\section{REST query}{

POST (multipart/form-data): ndex.api.config$api$network$create$url
data: CXNetworkStream = data
}

