% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_groups.R
\name{ndex.delete.group}
\alias{ndex.delete.group}
\title{Delete Group}
\usage{
ndex.delete.group(ndexcon, groupId)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection link{ndex.connect}}

\item{groupId}{character; unique ID (UUID) of the group}
}
\value{
NULL if successfull, else an error is thrown
}
\description{
Delete the group specified by groupId
}
\note{
Requires an authorized user! (ndex.connect with credentials)

Compatible to NDEx server version 2.0
}
\section{REST query}{

DELETE: ndex.conf$api$group$delete
}

\examples{
## Establish a server connection with credentials 
# ndexcon = ndex.connect('MyAccountName', 'MyPassword')
## Find user and get its id
# user = ndex.find.user.byName(ndexcon, 'MyAccountName')
# userId = user$externalId
## Find the user's groups and get one group id
# groups = ndex.user.list.groups(ndexcon, userId)
# groupId = groups[1,"externalId"]
#ndexr.delete.group(ndexcon,groupId)
NULL
}
