% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks_simple_operations.r
\name{ndex.delete.network}
\alias{ndex.delete.network}
\title{Delete a network}
\usage{
ndex.delete.network(ndexcon, networkId)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection link{ndex.connect}}

\item{networkId}{unique ID of the network}
}
\value{
NULL on success; Error else
}
\description{
Delete a network
}
\note{
Requires an authorized user! (ndex.connect with credentials)

Compatible to NDEx server version 1.3 and 2.0
}
\section{REST query}{

DELETE: ndex.conf$api$network$delete
}

\examples{
## Establish a server connections with credentials 
# ndexcon = ndex.connect('MyAccountName', 'MyPassword')
## Find a network and get its UUID
# networks = ndex.find.networks(ndexcon,"p53")
# networkId = networks[1,"externalId"] 
## Delete the network
# ndex.delete.network(ndexcon, networkId)
NULL
}
