% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks_simple_operations.r
\name{ndex.delete.network}
\alias{ndex.delete.network}
\title{Delete a network}
\usage{
ndex.delete.network(ndexcon, networkId)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{networkId}{unique ID of the network}
}
\value{
NULL on success; Error else
}
\description{
Delete a network
}
\note{
Compatible to NDEx server version 1.3 and 2.0
}
\section{REST query}{

DELETE: ndex.api.config$api$network$delete
}

\examples{
\dontrun{
ndexcon = ndex.connect('MyAccountName', 'MyPassword', verbose=T)
pws = ndex.find.networks(ndexcon,"SomeNetwork", accountName='MyAccountName')
response = ndex.delete.network(ndexcon, pws[1,"externalId")}
}
