% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_users.R
\name{ndex.delete.user}
\alias{ndex.delete.user}
\title{Delete User}
\usage{
ndex.delete.user(ndexcon, userId)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection link{ndex.connect}}

\item{userId}{character; unique ID (UUID) of the user}
}
\value{
NULL if successfull, else an error is thrown
}
\description{
Deletes the authenticated user, removing any other objects in the database that depend on the user
}
\note{
Requires an authorized user! (ndex.connect with credentials)

Compatible to NDEx server version 2.0
}
\section{REST query}{

GET: ndex.conf$api$user$delete
}

\examples{
## Establish a server connection with credentials 
# ndexcon = ndex.connect('MyAccountName', 'MyPassword')
## Find user and get its id
# user = ndex.find.user.byName(ndexcon, 'SomeUserName')
# userId = user$externalId
## Delete user
# ndexr.delete.user(ndexcon, userId)
NULL
}
