% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ndex.find.networks}
\alias{ndex.find.networks}
\title{Search networks in NDEx (by description)
    network    POST    /network/search/{skipBlocks}/{blockSize}    SimpleNetworkQuery    NetworkSummary[]}
\usage{
ndex.find.networks(searchString = "", accountName, skipBlocks = 0,
  blockSize = 10)
}
\arguments{
\item{searchString}{string by which to search}

\item{accountName}{string; constrain search to networks administered by this account}

\item{skip}{how many networks to skip}

\item{top}{how many networks to show}
}
\value{
Data frame with network information: ID, name, whether it is public, edge and node count; source and format of network
}
\description{
Search networks in NDEx (by description)
    network    POST    /network/search/{skipBlocks}/{blockSize}    SimpleNetworkQuery    NetworkSummary[]
}
\note{
Search strings may be structured
}
\examples{
\dontrun{ndex.find.networks("calmodulin")}
}

