% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_users.R
\name{ndex.find.user.byId}
\alias{ndex.find.user.byId}
\title{Get User By UUID}
\usage{
ndex.find.user.byId(ndexcon, userId)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection link{ndex.connect}}

\item{userId}{character; unique ID (UUID) of the user}
}
\value{
list of properties describing the user (externalId, emailAddress, website, etc.). Throws error (404) if user isn't found!
}
\description{
Get User By UUID
}
\note{
Compatible to NDEx server version 2.0
}
\section{REST query}{

GET: ndex.conf$api$user$get$byId
}

\examples{
## Establish a server connection
ndexcon = ndex.connect()
## Find user by name
user = ndex.find.user.byName(ndexcon, 'ndextutorials')
## Find user by Id
user = ndex.find.user.byId(ndexcon, user$externalId)
}
