% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_users.R
\name{ndex.find.user.byName}
\alias{ndex.find.user.byName}
\title{Get User By Name}
\usage{
ndex.find.user.byName(ndexcon, name)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{name}{name of the user}
}
\value{
result
}
\description{
Get User By Name
}
\note{
Compatible to NDEx server version 1.3 and 2.0
}
\section{REST query}{

GET: ndex.api.config$api$user$get$byName
}

\examples{

}
