% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ndex.get.complete.network}
\alias{ndex.get.complete.network}
\title{Get complete network}
\usage{
ndex.get.complete.network(network_id)
}
\arguments{
\item{network_id}{unique ID of the network}
}
\value{
\code{\link{ndexgraph}} object
}
\description{
Get complete network
}
\details{
Uses getEdges (this procedure will return complete network with all elements)
Nodes use primary ID of the base term ('represents' element)
Edges use primary ID of the base term ('predicate', or 'p' element)
Mapping table for the nodes is retrieved ('alias' and 'related' terms) to facilitate conversions/data mapping
}
\section{REST query}{

This function runs GET query /network/{networkUUID}/asNetwork (returns Network object
}

