% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks.r
\name{ndex.get.limited.aspect}
\alias{ndex.get.limited.aspect}
\title{Get Some Aspect Elements from a Network as CX}
\usage{
ndex.get.limited.aspect(ndexcon, network_id, aspectName, limit)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{network_id}{unique ID of the network}

\item{aspectName}{string name of the aspect to retrieve}

\item{limit}{integer limiting the amount of output to retrieve}
}
\value{
\code{\link{RCX}} object
}
\description{
GET : /network/{networkId}/aspect/{aspectName}/{limit}
This method returns elements in the specified aspect up to the specified limit.
}
\section{REST query}{

GET : /network/{networkId}/aspect/{aspectName}/{limit}
}

