% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks.r
\name{ndex.get.network.metadata}
\alias{ndex.get.network.metadata}
\title{Get the Metadata Associated with a Network UUID}
\usage{
ndex.get.network.metadata(ndexcon, network_id)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{network_id}{unique ID of the network}
}
\value{
data.frame listing network metadata: ID, name, whether it is public, edge and node count; source and format of network
}
\description{
This function retrieves the metadata associated with the supplied network UUID.
}
\section{REST query}{

This function runs GET query /network/{networkId}/metadata    and returns the network metadata as a data.frame
}
\examples{
\dontrun{
ndexcon = ndex.connect(verbose=T)
pws = ndex.find.networks(ndexcon,"p53")
ndex.get.network.metadata(ndexcon,pws[1,"externalId"]) }
}

