% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks.r
\name{ndex.get.network.namespace}
\alias{ndex.get.network.namespace}
\title{Get the BEL Namespaces Associated with a Network UUID
GET : /network/{networkId}/namespace}
\usage{
ndex.get.network.namespace(ndexcon, network_id)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{network_id}{unique ID of the network}
}
\value{
data.frame listing network namespace
}
\description{
Get the BEL Namespaces Associated with a Network UUID
GET : /network/{networkId}/namespace
}

