% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks_simple_operations.r
\name{ndex.get.network}
\alias{ndex.get.network}
\title{Get complete network}
\usage{
ndex.get.network(ndexcon, networkId)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{networkId}{unique ID of the network}
}
\value{
\code{\link{RCX}} object
}
\description{
Get complete network
}
\details{
Uses getEdges (this procedure will return complete network with all elements)
Nodes use primary ID of the base term ('represents' element)
Edges use primary ID of the base term ('predicate', or 'p' element)
Mapping table for the nodes is retrieved ('alias' and 'related' terms) to facilitate conversions/data mapping
}
\note{
Compatible to NDEx server version 1.3 and 2.0
}
\section{REST query}{

GET: ndex.api.config$api$network$get
}

\examples{
\dontrun{
ndexcon = ndex.connect(verbose=T)
pws = ndex.find.networks(ndexcon,"p53")
rcx = ndex.get.network(ndexcon,pws[1,"externalId"]) }
}
