% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks.r
\name{ndex.get.network.summary}
\alias{ndex.get.network.summary}
\title{Get NetworkSummary by Network UUID
GET    /network/{networkUUID}       NetworkSummary}
\usage{
ndex.get.network.summary(ndexcon, network_id)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{network_id}{unique ID of the network}
}
\value{
List of network metadata: ID, name, whether it is public, edge and node count; source and format of network
}
\description{
Get NetworkSummary by Network UUID
GET    /network/{networkUUID}       NetworkSummary
}

