% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks.r
\name{ndex.get.network.summary}
\alias{ndex.get.network.summary}
\title{Get NetworkSummary by Network UUID}
\usage{
ndex.get.network.summary(ndexcon, network_id)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{network_id}{unique ID of the network}
}
\value{
List of network metadata: ID, name, whether it is public, edge and node count; source and format of network
}
\description{
This function retrieves the summary of the network identified by the supplied network UUID string.
}
\section{REST query}{

This function runs GET query /network/{networkUUID}    and returns a single NetworkSummary
}
\examples{
\dontrun{
ndexcon = ndex.connect(verbose=T)
pws = ndex.find.networks(ndexcon,"p53")
ndex.get.network.summary(ndexcon,pws[1,"externalId"]) }
}

