% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_groups.R
\name{ndex.group.delete.membership}
\alias{ndex.group.delete.membership}
\title{Remove a Group Member}
\usage{
ndex.group.delete.membership(ndexcon, groupId, userId)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection link{ndex.connect}}

\item{groupId}{character; unique ID (UUID) of the group}

\item{userId}{character; unique ID (UUID) of the user}
}
\value{
Empty string ("") on success, else error
}
\description{
Removes the member from the group
}
\note{
Requires an authorized user! (ndex.connect with credentials)

Compatible to NDEx server version 2.0
}
\section{REST query}{

DELETE: ndex.conf$api$user$membership$delete
}

\examples{
## Establish a server connection with credentials 
# ndexcon = ndex.connect('MyAccountName', 'MyPassword')
## Find user and get own id
# user = ndex.find.user.byName(ndexcon, 'MyAccountName')
# userId = user$externalId
## Find own groups and get one group id
# groups = ndex.user.list.groups(ndexcon, userId)
# groupId = groups[1,"externalId"]
## Find an other user of the group and get the id
# users = ndex.group.list.users(ndexcon, groupId)
## Choose one user 
# userId = users[1,"externalId"]
## Remove user from the group
# ndex.group.delete.membership(ndexcon, groupId, userId)
NULL
}
