% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks_aspects_and_metadata.r
\name{ndex.network.aspect.get.metadata}
\alias{ndex.network.aspect.get.metadata}
\title{Get the Metadata Associated with a Network UUID}
\usage{
ndex.network.aspect.get.metadata(ndexcon, networkId, aspect)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection link{ndex.connect}}

\item{networkId}{character; unique ID (UUID) of the network}

\item{aspect}{character; aspect name}
}
\value{
metadata for an aspect as list: consistencyGroup, elementCount, lastUpdate, data, name, properties, version and idCounter
}
\description{
This function retrieves the metadata associated with the supplied network UUID.
}
\note{
Compatible to NDEx server version 2.0

Server error (version 2.0) since March 13th 2017
}
\section{REST query}{

GET: ndex.conf$api$network$aspect$getMetaDataByName
}

\examples{
## Establish a server connection
ndexcon = ndex.connect()
## Find a network and get its UUID
networks = ndex.find.networks(ndexcon,"p53")
networkId = networks[1,"externalId"]
## Get the meta-data of an aspect of a network
ndex.network.aspect.get.metadata(ndexcon, networkId, 'nodeAttributes') 
}
