% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks_aspects_and_metadata.r
\name{ndex.network.aspect.get.metadata}
\alias{ndex.network.aspect.get.metadata}
\title{Get the Metadata Associated with a Network UUID}
\usage{
ndex.network.aspect.get.metadata(ndexcon, nuuid, aspect)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{nuuid}{unique ID of the network}
}
\value{
metadata for an aspect as list: consistencyGroup, elementCount, lastUpdate, data, name, properties, version and idCounter
}
\description{
This function retrieves the metadata associated with the supplied network UUID.
}
\note{
Compatible to NDEx server version 2.0

Server error (version 2.0) since March 13th 2017
}
\section{REST query}{

GET: ndex.api.config$api$network$aspect$getMetaDataByName
}
\examples{
\dontrun{
ndexcon = ndex.connect(verbose=T)
pws = ndex.find.networks(ndexcon,"p53")
ndex.network.aspect.get.metadata(ndexcon,pws[1,"externalId"]) }
}

