% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks_aspects_and_metadata.r
\name{ndex.network.get.aspect}
\alias{ndex.network.get.aspect}
\title{Get a Network Aspect As CX}
\usage{
ndex.network.get.aspect(ndexcon, network_id)
}
\description{
Get a Network Aspect As CX
}

