% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks_aspects_and_metadata.r
\name{ndex.network.get.aspect}
\alias{ndex.network.get.aspect}
\title{Get a Network Aspect As CX}
\usage{
ndex.network.get.aspect(ndexcon, networkId, aspect, size)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection link{ndex.connect}}

\item{networkId}{character; unique ID of the network}

\item{aspect}{character; name of the aspect}

\item{size}{integer; specifies the number of elements returned}
}
\value{
data.frame of the aspect data (the same as rcx[[aspectName]])
}
\description{
This function retrieves the provided aspect as CX. The result is the same as accessing an aspect of a RCX object.
}
\note{
Compatible to NDEx server version 1.3 and 2.0, but doesn't work for version 1.3
}
\section{REST query}{

GET: ndex.conf$api$network$aspect$getMetaDataByName
}

\examples{
## Establish a server connection
ndexcon = ndex.connect()
## Find a network and get its UUID
networks = ndex.find.networks(ndexcon,"p53")
networkId = networks[1,"externalId"]
## Get the aspect of a network
aspect = ndex.network.get.aspect(ndexcon, networkId, 'nodeAttributes')
aspect = ndex.network.get.aspect(ndexcon, networkId, 'nodeAttributes', 10)    # limit the returned elemets of the aspect to the first 10 elements
}
