% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks_ndex_properties.r
\name{ndex.network.get.permission}
\alias{ndex.network.get.permission}
\title{Get All Permissions on a Network}
\usage{
ndex.network.get.permission(ndexcon, network_id)
}
\description{
Get All Permissions on a Network
}

