% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks_provenance.R
\name{ndex.network.set.provenance}
\alias{ndex.network.set.provenance}
\title{Set Network Provenance}
\usage{
ndex.network.set.provenance(ndexcon, nuuid, provenance)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{nuuid}{unique ID of the network}
}
\value{
List of network metadata: ID, name, whether it is public, edge and node count; source and format of network
}
\description{
Updates the 'provenance' field of the network specified by 'networkId' to be the ProvenanceEntity object in the PUT data. The ProvenanceEntity object is expected to represent the current state of the network and to contain a tree-structure of ProvenanceEvent and ProvenanceEntity objects that describe the networks provenance history.
}
\note{
Compatible to NDEx server version 1.3 and 2.0
}
\section{REST query}{

GET: ndex.api.config$api$network$provenance$get
}

\examples{
\dontrun{
ndexcon = ndex.connect('MyAccountName', 'MyPassword', verbose=T)
networks = ndex.find.networks(ndexcon,"p53")
nuuid = networks[1,"externalId"]
provenance = ndex.network.set.provenance(ndexcon, nuuid, provenance) }
}
