% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks_ndex_properties.r
\name{ndex.network.set.systemProperties}
\alias{ndex.network.set.systemProperties}
\title{Set Network System Properties}
\usage{
ndex.network.set.systemProperties(ndexcon, network_id)
}
\description{
Set Network System Properties
}

