% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks_aspects_and_metadata.r
\name{ndex.network.update.aspect}
\alias{ndex.network.update.aspect}
\title{Update an Aspect of a Network}
\usage{
ndex.network.update.aspect(ndexcon, network_id)
}
\description{
Update an Aspect of a Network
}

