% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks_ndex_properties.r
\name{ndex.network.update.permission}
\alias{ndex.network.update.permission}
\title{Update Network Permission}
\usage{
ndex.network.update.permission(ndexcon, networkId, user = NULL,
  group = NULL, permission)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection link{ndex.connect}}

\item{networkId}{unique ID of the network}

\item{user}{character (optional); uuid of the user. Only either user or group may be set!}

\item{group}{character (optional); uuid of the group. Only either user or group may be set!}

\item{permission}{character (optional)("READ"|"WRITE"|"ADMIN"); type of permission to be given. If granted admin permission, the current admin loses the admin status.}
}
\value{
1 integer on success, 0 if user/group allready has this permissions on the network
}
\description{
Updates the permission of a user specified by userid or group specified by groupid for the network
}
\note{
Requires an authorized user! (ndex.connect with credentials)

Compatible to NDEx server version 1.3 and 2.0, but doesn't work for version 1.3

In version 1.3 the function only works for user permissions!
}
\section{REST query}{

GET: ndex.conf$api$network$permission$update
}

\examples{
## Establish a server connection with credentials 
# ndexcon = ndex.connect('MyAccountName', 'MyPassword')
## Find one of your networks and get its UUID
# networks = ndex.find.networks(ndexcon, accountName='MyAccountName')
# networkId = networks[1,"externalId"]
## Get the UUID for a user and group
# someUserUuid = "uuuuuuuu-ssss-eeee-rrrr-111111111111"
# someGroupUuid = "ggggggg-rrrr-oooo-uuuu-pppppppppppp"
## Change the permissions
# ndex.network.update.permission(ndexcon, networkId, user=someUserUuid, 'WRITE')
# ndex.network.update.permission(ndexcon, networkId, group=someGroupUuid, 'READ')
## Set a new admin (lose own admin status)
# ndex.network.update.permission(ndexcon, networkId, user=someUserUuid, 'ADMIN')
NULL
}
