% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks_ndex_properties.r
\name{ndex.network.update.profile}
\alias{ndex.network.update.profile}
\title{Update Network Profile}
\usage{
ndex.network.update.profile(ndexcon, nuuid)
}
\description{
Update Network Profile
}

