% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_RCX.r
\name{ndex.RCX2JSON}
\alias{ndex.RCX2JSON}
\title{Generate JSON data from RCX object}
\usage{
ndex.RCX2JSON(rcx, verbose = FALSE)
}
\arguments{
\item{rcx}{RCX object}

\item{verbose}{logical; whether to print out extended feedback}
}
\value{
json jsonlite json object if successfull, NULL otherwise
}
\description{
Generate JSON data from RCX object
}

