% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks.r
\name{ndex.set.network.readonly}
\alias{ndex.set.network.readonly}
\title{Set a Network Read-Only
GET : /network/{networkId}/setFlag/{parameter}={value}}
\usage{
ndex.set.network.readonly(ndexcon, network_id, readonly = c(TRUE, FALSE)[1])
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{network_id}{unique ID of the network}

\item{readonly}{logical, defaults to TRUE}
}
\value{
response from server
}
\description{
Set a Network Read-Only
GET : /network/{networkId}/setFlag/{parameter}={value}
}

