% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_networks_simple_operations.r
\name{ndex.update.network}
\alias{ndex.update.network}
\title{Update an Entire Network as CX}
\usage{
ndex.update.network(ndexcon, rcx, nuuid)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection}

\item{rcx}{\code{\link{RCX}} object}

\item{nuuid}{UUID of the network}
}
\value{
UUID of the updated network
}
\description{
This method updates/replaces a existing network on the NDEx server with new content.
}
\details{
This method updates a network on the NDEx server with new content from the given RCX object
}
\section{REST query}{

PUT (multipart/form-data): ndex.api.config$api$network$update$url
data: CXNetworkStream = data
}

