% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_RCXgraph.r
\name{ndex_internal_addAspects}
\alias{ndex_internal_addAspects}
\title{ndex_internal_addAspects}
\usage{
ndex_internal_addAspects(rcxgraph, rcx, verbose = FALSE)
}
\arguments{
\item{rcxgraph}{RCXgraph object}

\item{rcx}{RCX object}

\item{verbose}{logical; whether to print out extended feedback}
}
\value{
returns object of class RCXgraph
}
\description{
ndex_internal_addAspects
}
\examples{
NULL
}
\keyword{internal}
