% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_connect.r
\name{ndex_rest_PUT}
\alias{ndex_rest_PUT}
\title{Generic PUT query to API}
\usage{
ndex_rest_PUT(ndexcon, route, data = NULL, multipart = FALSE, raw = FALSE)
}
\arguments{
\item{ndexcon}{object of class NDEXConnection \code{\link{ndex.connect}}}

\item{route}{Character (route to specific REST query)}

\item{data}{Whatever data to be supplied with query. Should be valid JSON}

\item{multipart}{Whatever data to be supplied with query. Should be valid JSON}

\item{raw}{Specifies if server response should be returned in raw, or if jsonlite::fromJSON is called first. Defaults to FALSE.}
}
\value{
JSON response from REST server, NULL if no valid JSON was received. if parameter raw is TRUE, the raw response is returned without a call to jsonlite::fromJSON.
}
\description{
This functions is internal.
}
\details{
Simply execute HTTP PUT on URL host/route and fetch whatever data REST server returns 
Making sure the route is well-formed is the job of calling function
Making sure the data is well-formed is also the job of calling function
}
\examples{
\dontrun{
ndexcon = ndex.connect(verbose=T)
ndex_rest_PUT(ndexcon, "/networks/api", data)
ndex_rest_PUT(ndexcon, "/networks/api", data, raw=T)
ndex_rest_PUT(ndexcon, "/networks/api", list(some=data, other=data2), multipart=T)
}
}
\seealso{
\code{\link{ndex_rest_GET}},  \code{\link{ndex_rest_POST}},  \code{\link{ndex_rest_PUT}} and \code{\link{ndex_rest_DELETE}}
}

