\name{ndexr-package}
\alias{ndexr-package}
\alias{ndexr}
\docType{package}
\title{
NDEx R client library
}
\description{
\bold{~~ To be added}: More about what it does (maybe more than one line)
}
\details{
\tabular{ll}{
Package: \tab ndexr\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-06-05\cr
License: \tab TBD\cr
}
\bold{~~ To be added} An overview of how to use the package, including the most important functions
}
\author{
Alex Ishkin <aleksandr.ishkin@thomsonreuters.com>

Maintainer: Dexter Pratt <depratt@ucsc.edu>

}
\references{
\link[http://www.ndexbio.org/]{http://www.ndexbio.org/}
}
\keyword{ package }
\seealso{
\code{\link[RCurl:RCurl-package]{RCurl}}, \code{\link[igraph:igraph-package]{igraph}}, \code{\link[RJSONIO:RJSONIO-package]{RJSONIO}}
}
\examples{
\bold{~~ To be added}
}
