% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_RCX.r
\name{rcx.asNewNetwork}
\alias{rcx.asNewNetwork}
\title{Remove all interfering NDEx artefacts from RCX object}
\usage{
rcx.asNewNetwork(rcx)
}
\arguments{
\item{rcx}{RCX object}
}
\value{
\code{\link{RCX}} object
}
\description{
Remove all interfering NDEx artefacts from RCX object
}
\details{
After a RCX object is downloaded from an NDEx server, it will contain some aspects that are not present in a newly generated network, i.e. ndexStatus', provenanceHistory' and 'status'.
Removing those aspects might be useful in some cases.
}
\examples{
## Establish a server connection
ndexcon = ndex.connect()
## Find a network and get its UUID
networks = ndex.find.networks(ndexcon,"p53")
networkId = networks[1,"externalId"]
## Get the network data 
rcx = ndex.get.network(ndexcon, networkId)
## Remove NDEx artefacts
rcx = rcx.asNewNetwork(rcx)
\dontrun{
rcxjson = rcx.toJSON(rcx)
ndex.create.network(ndexcon, rcxjson) 
}
}
\seealso{
\code{\link{rcx.fromJSON}} \code{\link{ndex.get.network}}
}
