% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_RCX.r
\name{rcx.aspect.toJSON}
\alias{rcx.aspect.toJSON}
\title{Generate JSON data for a single aspect of a RCX object}
\usage{
rcx.aspect.toJSON(rcxAspect, verbose = FALSE, pretty = FALSE)
}
\arguments{
\item{rcxAspect}{aspect in RCX object (rcx[[aspectName]])}

\item{verbose}{logical; whether to print out extended feedback}

\item{pretty}{logical; adds indentation whitespace to JSON output}
}
\value{
json object if successfull, empty string otherwise
}
\description{
Generate JSON data for a single aspect of a RCX object
}
\examples{
\dontrun{
ndexcon = ndex.connect(verbose=T)
pws = ndex.find.networks(ndexcon,"p53")
rcx = ndex.get.network(ndexcon,pws[1,"externalId"]) 
rcxNodesJson = rcx.aspect.toJSON(rcx$nodes) }
}
\seealso{
\code{\link{rcx.toJSON}} and \code{\link{rcx.fromJSON}}
}
