% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_ngraph.r
\name{rcx.toNGraph}
\alias{rcx.toNGraph}
\title{Create ngraph object from RCX object}
\usage{
rcx.toNGraph(rcx, verbose = FALSE)
}
\arguments{
\item{rcx}{RCX object}

\item{verbose}{logical; whether to print out extended feedback}
}
\value{
returns object of class ngraph if successfull, NULL otherwise
}
\description{
Create ngraph object from RCX object
}
\note{
Wrapper function for \code{\link{ngraph.fromRCX}}
}
\examples{
## Establish a server connection
ndexcon = ndex.connect()
## Find one of your networks and get its UUID
networks = ndex.find.networks(ndexcon)
networkId = networks[1,"externalId"]
## Get the network data 
rcx = ndex.get.network(ndexcon, networkId) 
## Convert to nGraph
ngraph = rcx.toNGraph(rcx) 
}
