% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_ngraph.r
\name{rcx.toNGraph}
\alias{rcx.toNGraph}
\title{Create ngraph object from RCX object}
\usage{
rcx.toNGraph(rcx, verbose = FALSE)
}
\description{
Create ngraph object from RCX object
}
\note{
Wrapper function for \code{\link{ngraph.fromRCX}}
}
