% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_RCX.r
\name{rcx_asNewNetwork}
\alias{rcx_asNewNetwork}
\title{Remove all interfering NDEx artefacts from RCX object}
\usage{
rcx_asNewNetwork(rcx)
}
\arguments{
\item{rcx}{RCX object}
}
\value{
\code{\link{RCX}} object
}
\description{
Remove all interfering NDEx artefacts from RCX object
}
\details{
After a RCX object is downloaded from an NDEx server, it will contain some aspects that are not present in a newly generated network, i.e. ndexStatus', provenanceHistory' and 'status'.
Removing those aspects might be useful in some cases.
}
\examples{
## Create an RCX object
rcx = rcx_new(c('@id'=1, n='Some Name', r='HGNC:Symbol'))
## Remove NDEx artefacts
rcx = rcx_asNewNetwork(rcx)
\dontrun{
rcxjson = rcx_toJSON(rcx)
ndex_create_network(ndexcon, rcxjson)
}
}
\seealso{
\code{\link{rcx_fromJSON}}
}
