% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_RCX.r
\name{rcx_aspect_toJSON}
\alias{rcx_aspect_toJSON}
\title{Generate JSON data for a single aspect of a RCX object}
\usage{
rcx_aspect_toJSON(rcxAspect, verbose = FALSE, pretty = FALSE)
}
\arguments{
\item{rcxAspect}{aspect in RCX object (rcx[[aspectName]])}

\item{verbose}{logical; whether to print out extended feedback}

\item{pretty}{logical; adds indentation whitespace to JSON output}
}
\value{
json object if successfull, empty string otherwise
}
\description{
\strong{Note: In future `ndexr` uses the \link[RCX]{RCX-object} from the corresponding package to handle the networks!}
\strong{For a replacement of this function see \link[RCX]{rcxToJson}, \link[RCX]{writeCX} and \link[RCX]{toCX}!}
}
\examples{
\dontrun{
## Create an RCX object
rcx = rcx_new(c('@id'=1, n='Some Name', r='HGNC:Symbol'))
## Convert RCX aspect to JSON
rcxNodesJson = ndexr:::rcx_aspect_toJSON(rcx$nodes)
}
}
\seealso{
\code{\link{rcx_toJSON}} and \code{\link{rcx_fromJSON}}
}
\keyword{internal}
