% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_RCX.r
\name{rcx_toJSON}
\alias{rcx_toJSON}
\title{Generate JSON data from RCX object}
\usage{
rcx_toJSON(rcx, verbose = FALSE, pretty = FALSE)
}
\arguments{
\item{rcx}{RCX object}

\item{verbose}{logical; whether to print out extended feedback}

\item{pretty}{logical; adds indentation whitespace to JSON output}
}
\value{
json jsonlite json object if successfull, NULL otherwise
}
\description{
Generate JSON data from RCX object
}
\examples{
## Establish a server connection
ndexcon = ndex_connect()
## Find a network and get its UUID
networks = ndex_find_networks(ndexcon,"p53")
networkId = networks[1,"externalId"]
## Get the network data 
rcx = ndex_get_network(ndexcon, networkId)
## Convert RCX to JSON
rcxjson = rcx_toJSON(rcx)
}
\seealso{
\code{\link{ngraph_fromRCX}} \code{\link{ngraph_toRCX}} \code{\link{rcx_fromJSON}}
}
