\name{validate_ndexgraph}
\alias{validate_ndexgraph}
\title{Validator for ndexgraph class}
\usage{
  validate_ndexgraph(object)
}
\arguments{
  \item{object}{R object to be tested}
}
\value{
  returns TRUE if the object is valid instance of
  \code{ndexgraph} S4 class; otherwise returns vector with
  error messages
}
\description{
  Validator for ndexgraph class
}
\note{
  Additional validations may be needed!
}

